Adicionando melhores colisões e criando cenario alpha a partir de tileset

+++ engine/src/components/image.cpp
@@ -51,8 +51,8 @@ bool ImageComponent::shutdown()
 void ImageComponent::draw()
 {
     SDL_Rect renderQuad = {
-        m_game_object->position.getX(),
-        m_game_object->position.getY(),
+        (int) m_game_object->position.getX(),
+        (int) m_game_object->position.getY(),
         m_game_object->w,
         m_game_object->h
     };
